// src/api/models/CalendarConnectionRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CalendarConnectionRead = {
    name: string;
    color: string;
    url: string;
    provider?: (string | null);
    enabled?: boolean;
    id: number;
    profile_id: number;
    created_date: string;
    updated_date?: (string | null);
    last_fetched_at?: (string | null);
    last_error?: (string | null);
};